GCC 3.4.3 complains about float vector types not having a float base type. 	(The info docs said the base type was ignored, but I guess not any more)

+++ simd.h
@@ -220,15 +220,15 @@ template <typename T> inline int32_t sround (T op) { fround<T,int32_t> obj; retu
 #else
 #define VECTOR_ATTRIBUTE(x)
 #endif
-typedef int v8qi_t VECTOR_ATTRIBUTE (V8QI);
-typedef int v4hi_t VECTOR_ATTRIBUTE (V4HI);
-typedef int v8hi_t VECTOR_ATTRIBUTE (V8HI);
-typedef int v2si_t VECTOR_ATTRIBUTE (V2SI);
-typedef int v4si_t VECTOR_ATTRIBUTE (V4SI);
-typedef int v1di_t VECTOR_ATTRIBUTE (V1DI);
-typedef int v2sf_t VECTOR_ATTRIBUTE (V2SF);
-typedef int v4sf_t VECTOR_ATTRIBUTE (V4SF);
-typedef int v2df_t VECTOR_ATTRIBUTE (V2DF);
+typedef uint8_t v8qi_t VECTOR_ATTRIBUTE (V8QI);
+typedef uint16_t v4hi_t VECTOR_ATTRIBUTE (V4HI);
+typedef uint16_t v8hi_t VECTOR_ATTRIBUTE (V8HI);
+typedef uint32_t v2si_t VECTOR_ATTRIBUTE (V2SI);
+typedef uint32_t v4si_t VECTOR_ATTRIBUTE (V4SI);
+typedef uint32_t v1di_t VECTOR_ATTRIBUTE (V1DI);
+typedef float v2sf_t VECTOR_ATTRIBUTE (V2SF);
+typedef float v4sf_t VECTOR_ATTRIBUTE (V4SF);
+typedef double v2df_t VECTOR_ATTRIBUTE (V2DF);
 #undef VECTOR_ATTRIBUTE
 
 //----------------------------------------------------------------------
